Removes code duplicacy
