+++ stores-service-module/store-common-module/src/test/java/com/kgcorner/topspin/dtos/StoreDTOTest.java
@@ -1,5 +1,6 @@
 package com.kgcorner.topspin.dtos;
 
+import com.kgcorner.topspin.model.AbstractStore;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
@@ -17,167 +18,23 @@ import java.util.List;
 public class StoreDTOTest {
 
     private StoreDTO storeDTO;
-
+    private AbstractStore store;
     @Before
     public void setUp() {
-        storeDTO = new StoreDTO();
+        storeDTO = new StoreDTO(store);
     }
 
-    @Test
-    public void setStoreId() {
-        String storeId = "storeID";
-        storeDTO.setStoreId(storeId);
-        Assert.assertEquals(storeId, storeDTO.getStoreId());
-    }
 
-    @Test
-    public void setName() {
-        String name = "Amazon";
-        storeDTO.setName(name);
-        Assert.assertEquals(name, storeDTO.getName());
-    }
-
-    @Test
-    public void setLink() {
-        String link = "link";
-        storeDTO.setLink(link);
-        Assert.assertEquals(link, storeDTO.getLink());
-    }
-
-    @Test
-    public void setLogo() {
-        String logo = "logo";
-        storeDTO.setLogo(logo);
-        Assert.assertEquals(logo, storeDTO.getLogo());
-    }
-
-    @Test
-    public void setOpenOut() {
-        boolean openOut = true;
-        storeDTO.setOpenOut(openOut);
-        Assert.assertTrue(storeDTO.isOpenOut());
-    }
-
-    @Test
-    public void setAffiliated() {
-        boolean affiliated =true;
-        storeDTO.setAffiliated(affiliated);
-        Assert.assertTrue(storeDTO.isAffiliated());
-    }
-
-    @Test
-    public void setPlaceHolder() {
-        String placeHolder = "placeHolder";
-        storeDTO.setPlaceHolder(placeHolder);
-        Assert.assertEquals(placeHolder, storeDTO.getPlaceHolder());
-    }
-
-    @Test
-    public void setSurferPlaceHolder() {
-        String SurferPlaceHolder = "SurferPlaceHolder";
-        storeDTO.setSurferPlaceHolder(SurferPlaceHolder);
-        Assert.assertEquals(SurferPlaceHolder, storeDTO.getSurferPlaceHolder());
-    }
-
-    @Test
-    public void setSearchUrl() {
-        String searchUrl = "searchUrl";
-        storeDTO.setSearchUrl(searchUrl);
-        Assert.assertEquals(searchUrl, storeDTO.getSearchUrl());
-    }
-
-    @Test
-    public void setFixedUrls() {
-        String fixedUrl = "fixedUrl";
-        storeDTO.setFixedUrls(fixedUrl);
-        Assert.assertEquals(fixedUrl, storeDTO.getFixedUrls());
-    }
-
-    @Test
-    public void setPidRegex() {
-        String regex = "regex";
-        storeDTO.setPidRegex(regex);
-        Assert.assertEquals(regex, storeDTO.getPidRegex());
-    }
-
-    @Test
-    public void setGender() {
-        String gender = "Male";
-        storeDTO.setGender(gender);
-        Assert.assertEquals(gender, storeDTO.getGender());
-    }
-
-    @Test
-    public void setAffiliateId() {
-        String affiliateId = "affiliateId";
-        storeDTO.setAffiliateId(affiliateId);
-        Assert.assertEquals(affiliateId, storeDTO.getAffiliateId());
-    }
-
-    @Test
-    public void setMaxCashback() {
-        String cashback = "50%";
-        storeDTO.setMaxCashback(cashback);
-        Assert.assertEquals(cashback, storeDTO.getMaxCashback());
-    }
-
-    @Test
-    public void setDescription() {
-        String description = "description";
-        storeDTO.setDescription(description);
-        Assert.assertEquals(description, storeDTO.getDescription());
-    }
-
-    @Test
-    public void setLongDescription() {
-        String longDescription = "longDescription";
-        storeDTO.setLongDescription(longDescription);
-        Assert.assertEquals(longDescription, storeDTO.getLongDescription());
-    }
-
-    @Test
-    public void setBannerImage() {
-        String bannerImage = "bannerImage";
-        storeDTO.setBannerImage(bannerImage);
-        Assert.assertEquals(bannerImage, storeDTO.getBannerImage());
-    }
-
-    @Test
-    public void setThumbnailImage() {
-        String thumbNailImage = "thumbNailImage";
-        storeDTO.setThumbnailImage(thumbNailImage);
-        storeDTO.setThumbnailImage(thumbNailImage);
-        Assert.assertEquals(thumbNailImage, storeDTO.getThumbnailImage());
-    }
-
-    @Test
-    public void setLargeImage() {
-        String largeImage = "largeImage";
-        storeDTO.setLargeImage(largeImage);
-        Assert.assertEquals(largeImage, storeDTO.getLargeImage());
-    }
-
-    @Test
-    public void setTagLine() {
-        String tagLine = "tagLine";
-        storeDTO.setTagLine(tagLine);
-        Assert.assertEquals(tagLine, storeDTO.getTagLine());
-    }
-
-    @Test
-    public void setActive() {
-        boolean active = true;
-        storeDTO.setActive(active);
-        Assert.assertTrue(storeDTO.isActive());
-    }
 
     @Test
     public void setCategories() {
         List<CategoryDTO> categoryDTOS = new ArrayList<>();
         for (int i = 0; i < 10; i++) {
-            categoryDTOS.add(new CategoryDTO());
+            CategoryDTO categoryDTO = new CategoryDTO(null);
+            categoryDTOS.add(categoryDTO);
+            storeDTO.addCategories(categoryDTO);
         }
-        storeDTO.setCategories(categoryDTOS);
+
         Assert.assertEquals(categoryDTOS, storeDTO.getCategories());
     }
 }
