+++ stores-service-module/store-service/src/test/java/com/kgcorner/topspin/service/CategoryServiceTest.java
@@ -1,6 +1,7 @@
 package com.kgcorner.topspin.service;
 
 import com.kgcorner.topspin.dtos.CategoryDTO;
+import com.kgcorner.topspin.model.AbstractCategory;
 import com.kgcorner.topspin.model.Category;
 import com.kgcorner.topspin.model.factory.CategoryFactory;
 import com.kgcorner.topspin.persistence.CategoryPersistenceLayer;
@@ -11,6 +12,7 @@ import org.powermock.api.mockito.PowerMockito;
 import org.powermock.reflect.Whitebox;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
 
@@ -24,10 +26,12 @@ public class CategoryServiceTest {
     private CategoryFactory categoryFactory;
     private CategoryPersistenceLayer persistenceLayer;
     private CategoryService categoryService;
+    private DemoCategory demoCategory;
 
     @Before
     public void setUp() throws Exception {
         categoryService = new CategoryService();
+        demoCategory = new DemoCategory();
         categoryFactory = PowerMockito.mock(CategoryFactory.class);
         persistenceLayer = PowerMockito.mock(CategoryPersistenceLayer.class);
         Whitebox.setInternalState(categoryService, "categoryFactory", categoryFactory);
@@ -36,8 +40,8 @@ public class CategoryServiceTest {
 
     @Test
     public void getCategory() {
-        CategoryDTO category = new CategoryDTO();
-        category.setName("name");
+        CategoryDTO category = new CategoryDTO(demoCategory);
+        demoCategory.setName("name");
         String id = "id";
         PowerMockito.when(persistenceLayer.getCategory(id)).thenReturn(category);
         Category result = categoryService.getCategory(id);
@@ -50,9 +54,9 @@ public class CategoryServiceTest {
     public void createCategory() {
         String name = "name";
         String description = "description";
-        CategoryDTO category = new CategoryDTO();
-        category.setName(name);
-        category.setDescription(description);
+        CategoryDTO category = new CategoryDTO(demoCategory);
+        demoCategory.setName(name);
+        demoCategory.setDescription(description);
         PowerMockito.when(categoryFactory.createCategoryModel(name, description)).thenReturn(category);
         PowerMockito.when(persistenceLayer.createCategory(category)).thenReturn(category);
         Category result = categoryService.createCategory(name, description);
@@ -66,9 +70,9 @@ public class CategoryServiceTest {
     public void updateCategory() {
         String name = "name";
         String description = "description";
-        CategoryDTO category = new CategoryDTO();
-        category.setName(name);
-        category.setDescription(description);
+        CategoryDTO category = new CategoryDTO(demoCategory);
+        demoCategory.setName(name);
+        demoCategory.setDescription(description);
         String id = "id";
         PowerMockito.when(categoryFactory.createCategoryModel(name, description)).thenReturn(category);
         PowerMockito.when(persistenceLayer.getCategory(id)).thenReturn(category);
@@ -85,10 +89,27 @@ public class CategoryServiceTest {
         int page = 1;
         int maxItem = 10;
         List<Category> categories = new ArrayList<>();
-        categories.add(new CategoryDTO());
+        categories.add(new CategoryDTO(demoCategory));
         PowerMockito.when(persistenceLayer.getAllCategories(page, maxItem)).thenReturn(categories);
         List<CategoryDTO> result = categoryService.getAllCategories(page, maxItem);
         Assert.assertNotNull(result);
         Assert.assertEquals(categories.size(), result.size());
+    }
+
+    class DemoCategory extends AbstractCategory {
+        private String cateId;
+
+        public void setCategoryId(String cateId) {
+            this.cateId = cateId;
+        }
+        @Override
+        public String getCategoryId() {
+            return cateId;
+        }
+
+        @Override
+        public List<? extends Category> getChildren() {
+            return Collections.emptyList();
+        }
     }
 }
